WEB/concesionario/requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0
django-import-export==2.5.0
django-suit==2.0a1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
